Update of the 3D view properly. Fix area selection when hovering out
The BabylonJS scene wasn't updated properly because of a simple bug in
the `updateScene` function - we iterate over the meshes via an index,
and remove meshes in the loop - which causes the mesh indices to
change, so we don't always clear the whole scene properly.

Fix the interaction between area selection and hovering out of the
PIXI.JS container - previously if you hover out you kept
area-selecting, even if you release the left mouse button. This is now
fixed by tracking when the mouse is inside the container